DB migration, edit user profile

+++ src/main/java/hello/domain/Message.java
@@ -6,7 +6,7 @@ import javax.persistence.*;
 public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @Column
     private String text;
@@ -37,7 +37,7 @@ public class Message {
         this.author = author;
     }
 
-    public Message(Integer id) {
+    public Message(Long id) {
         this.id = id;
     }
 
@@ -63,11 +63,11 @@ public class Message {
         this.tag = tag;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
